separate dataset preperation
