--- core/config/abc123
@@ -1,59 +1,0 @@
-async_upload = true
-bucket_no_upscale = true
-bucket_reso_steps = 32
-cache_latents = true
-cache_latents_to_disk = true
-caption_extension = ".txt"
-clip_skip = 1
-dynamo_backend = "no"
-enable_bucket = true
-epoch = 10
-gradient_accumulation_steps = 1
-gradient_checkpointing = true
-huber_c = 0.1
-huber_schedule = "snr"
-huggingface_path_in_repo = "test"
-huggingface_repo_id = "diagonalge/diffusiontest"
-huggingface_repo_type = "model"
-huggingface_repo_visibility = "public"
-huggingface_token = ""
-learning_rate = 1.0
-loss_type = "l2"
-lr_scheduler = "constant"
-lr_scheduler_args = []
-lr_scheduler_num_cycles = 1
-lr_scheduler_power = 1
-max_bucket_reso = 2048
-max_data_loader_n_workers = 0
-max_grad_norm = 1
-max_timestep = 1000
-max_token_length = 75
-max_train_steps = 10
-min_bucket_reso = 256
-min_snr_gamma = 5
-mixed_precision = "bf16"
-network_alpha = 16
-network_args = []
-network_dim = 32
-network_module = "networks.lora"
-no_half_vae = true
-noise_offset_type = "Original"
-optimizer_args = [ "weight_decay=0.01", "decouple=True", "d0=0.0001", "use_bias_correction=True",]
-optimizer_type = "Prodigy"
-output_dir = "/app/outputs"
-output_name = "kohyatest"
-pretrained_model_name_or_path = "SG161222/RealVisXL_V4.0"
-prior_loss_weight = 1
-resolution = "1024,1024"
-sample_prompts = ""
-sample_sampler = "euler_a"
-save_every_n_epochs = 2
-save_model_as = "safetensors"
-save_precision = "bf16"
-scale_weight_norms = 5
-text_encoder_lr = 1
-train_batch_size = 4
-train_data_dir = "./core/dataset/images/abc123/img/"
-training_comment = ""
-unet_lr = 1
-xformers = true
